+++ src/modules/auth/components/EmailInput/index.ts
@@ -1,0 +1,1 @@
+export * from './EmailInput'
