+++ src/modules/auth/components/LoginForm/index.ts
@@ -1,0 +1,1 @@
+export * from './LoginForm'
