+++ src/modules/auth/components/PasswordInput/index.ts
@@ -1,0 +1,1 @@
+export * from './PasswordInput'
